requirements.txt
django
django-rest-framework
python-decouple
psycopg2-binary
scapy